requirement.txt
Flask==2.0.2
Flask-Cors==3.0.9
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Mako==1.3.5
MarkupSafe==2.0.1
Pillow==9.5.0
PyMySQL==1.1.0
SQLAlchemy==1.4.25
Werkzeug==2.0.2
alembic==1.13.3
blinker==1.8.2
click==8.1.7
greenlet==3.1.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
pip==24.2
setuptools==41.2.0
six==1.16.0
typing-extensions==4.12.2
wtforms==3.1.2
zipp==3.20.2
scikit-learn==1.2.2
jieba==0.42.1
numpy==1.23.5
urllib3==2.0.2
wordcloud==1.9.4